requirements.txt
Flask==3.0.3
PyJWT==2.8.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
flask-cors==4.0.1
python-dotenv==1.0.1
